// frontend/src/types/enums.ts
export enum ClimateTypes {
    TROPICAL_SAVANNA = "TROPICAL_SAVANNA",
    OCEANIC = "OCEANIC",
    STEPPE = "STEPPE"
}

export enum GovernmentTypes {
    PUPPET_STATE = "PUPPET_STATE",
    OLIGARCHY = "OLIGARCHY",
    THALASSOCRACY = "THALASSOCRACY",
    THEOCRACY= "THEOCRACY",
}

export enum LivingStandardTypes {
    ULTRA_HIGH = "ULTRA_HIGH",
    MEDIUM = "MEDIUM",
    LOW = "LOW",
    VERY_LOW = "VERY_LOW",
    NIGHTMARE = "NIGHTMARE"
}